perseus: turn off compressed apex
Google introduced compressed apex on android 12 [1], and the flag is
enabled by default in updatable_apex.mk [2]. Perseus cannot boot
compressed apex with encryption, so disable it.

ref:
[1] https://source.android.com/devices/tech/ota/apex#compressed-apex
[2] https://android.googlesource.com/platform/build/+/4a1240de8a2ee5795a30df123058677d3999145e

basically, if this isn't disable, when the device is encrypted, they are causing bootloops.

As the only advantage is some space gain on disk, we can
simply disable them: keep the variable immediatly after including the
updatable_apex.mk to avoid being overwritten.

actually i can cherry-pick this, but whatever... wants to write details for perseus instead

Signed-off-by: lolipuru <banstellazz@gmail.com>